% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\docType{data}
\name{asset}
\alias{asset}
\title{Assets}
\format{
An object of class \code{data.frame} with 2157 rows and 6 columns.
}
\usage{
asset
}
\description{
A dataset containing investment, the indices of investment and
the price indices of investment
}
\details{
@format A data frame:
\describe{
\item{prv}{provinces}
\item{yr}{year}
\item{invest}{total fixed capital formation}
\item{InvestIndex}{index of fixed capital formation}
\item{InvestPrice}{price index of investment in fixed assets}
\item{depr}{depreciation}
}
}
\keyword{datasets}
